pyproject.toml
[project]
name = "hlsd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.12",
    "asyncio>=3.4.3",
    "m3u8>=6.0.0",
    "typed-ffmpeg>=2.6.1",
]
